All Link errors solved and reason understood!!!
