+++ .gitignore
@@ -40,3 +40,6 @@
 /GraphPrint/x64/Debug
 /Debug
 /GraphPrint/Debug
+/Output
+/GraphPrint/Output
+/Framework/Output
