+++ Framework/Himani/Himani/WinProc.h
@@ -74,92 +74,6 @@ struct winThunk {
 #endif
 
 
-#if defined(_M_IX86)
-#pragma pack(push, 1)
-
-bool winThunk::Init(void* Instance_to_attach, DWORD_PTR Function_to_suppply) {
-	m_push1 = 0x34ff; //ff 34 23 push DWORD PTR [esp]
-	m_push2 = 0xc724;
-
-	m_mov1 = 0x2444; //c7 44 24 04 mov dword ptr [esp +0x4],
-	m_mov2 = 0x04l;
-
-	m_this = PtrToUlong(Instance_to_attach);
-
-	m_jmp = 0xe9; //jmp
-
-				  //Calculate relative address of proc to jump to
-	m_relproc = unsigned long((INT_PTR)Function_to_suppply - ((INT_PTR)this + sizeof(winThunk)));
-
-	//write block from data cache and flush from instruction cache
-	if (FlushInstructionCache(GetCurrentProcess(), this, sizeof(winThunk)))
-		//Success
-		return true;
-	else
-		//Fail
-		return false;
-}
-
-WNDPROC winThunk::GetThunkAddress() {
-	return (WNDPROC)this;
-}
-
-void* winThunk::operator new(size_t, HANDLE HeapAddr) {
-	eHeapAddr = HeapAddr;
-
-	return HeapAlloc(HeapAddr, 0, sizeof(winThunk));
-}
-
-void winThunk::operator delete(void* pThunk) {
-	HeapFree(eHeapAddr, 0, pThunk);
-}
-
-HANDLE winThunk::eHeapAddr = NULL;
-#pragma pack(pop)
-
-#elif defined(_M_AMD64)
-
-#pragma pack(push, 2)
-
-bool winThunk::Init(void *pThis, DWORD_PTR proc) {
-	RaxMov = 0xb848;                    //movabs rax (48 B8), pThis
-	RaxImm = (unsigned long long)pThis; //
-	RspMov = 0x24448948;                //mov qword ptr [rsp+28h], rax (48 89 44 24 28)
-	RspMov1 = 0x9028;                    //to properly byte align the instruction we add a nop (no operation) (90)
-	Rax2Mov = 0xb848;                    //movabs rax (48 B8), proc
-	ProcImm = (unsigned long long)proc;
-	RaxJmp = 0xe0ff;                     //jmp rax (FF EO)
-	//write block from data cache and flush from instruction cache
-	if (FlushInstructionCache(GetCurrentProcess(), this, sizeof(winThunk)))
-		//Success
-		return true;
-	else
-		//Fail
-		return false;
-
-}
-
-WNDPROC winThunk::GetThunkAddress() {
-	return (WNDPROC)this;
-}
-
-void* winThunk::operator new(size_t, HANDLE HeapAddr) {
-	eHeapAddr = HeapAddr;
-
-	return HeapAlloc(HeapAddr, 0, sizeof(winThunk));
-}
-
-
-void winThunk::operator delete(void* pThunk) {
-	HeapFree(eHeapAddr, 0, pThunk);
-}
-
-HANDLE winThunk::eHeapAddr = NULL;
-
-#pragma pack(pop)
-
-#endif
-
 
 
 /////////////////////////////////////////////////////////////////////////////////////////////////////
@@ -169,7 +83,7 @@ HANDLE winThunk::eHeapAddr = NULL;
 
 
 
-namespace Himani{
+namespace Himani {
 
 
 	//The Default Callback Windows Procedure for every window
@@ -185,68 +99,71 @@ namespace Himani{
 
 #pragma region HBaseWinProc
 
-	class HBaseWinProc {
+	class FunctionThunk {
+		friend class HWindowsProc;
+		friend class HDialogProc;
 	public:
-		HBaseWinProc();
-		~HBaseWinProc();
+		FunctionThunk();
+		~FunctionThunk();
 
 	protected:
 		//Returns the windows Procedure specific to our class Object
 		WNDPROC Procedure() {
 			return thunk->GetThunkAddress();
 		}
-		virtual LRESULT MessageFunc(HWND hwnd, UINT message, WPARAM wParam, LPARAM lParam) {
-			return DefWindowProc(hwnd, message, wParam, lParam);
-		}
+
+
 	private:
 		winThunk * thunk;
 		static HANDLE eHeapAddr;
 		static int objInstances;
 
+	};
+
+	class HWindowsProc :public FunctionThunk {
+	public:
+		HWindowsProc() {
+			thunk->Init(this, (DWORD_PTR)WndProc);
+		}
+
+	protected:
+		virtual LRESULT MessageFunc(HWND hwnd, UINT message, WPARAM wParam, LPARAM lParam) {
+			return DefWindowProc(hwnd, message, wParam, lParam);
+		}
+
+	private:
 #if defined(_M_IX86)
-		static LRESULT CALLBACK WndProc(HBaseWinProc* pThis, HWND hwnd, UINT message, WPARAM wParam, LPARAM lParam) {
+		static LRESULT CALLBACK WndProc(HWindowsProc* pThis, HWND hwnd, UINT message, WPARAM wParam, LPARAM lParam) {
 #elif defined(_M_AMD64)
-		static LRESULT CALLBACK WndProc(HWND hwnd, UINT message, WPARAM wParam, LPARAM lParam, HBaseWinProc* pThis) {
+		static LRESULT CALLBACK WndProc(HWND hwnd, UINT message, WPARAM wParam, LPARAM lParam, HWindowsProc* pThis) {
 #endif
 			return pThis->MessageFunc(hwnd, message, wParam, lParam);
 		}
 
 	};
 
-	HBaseWinProc::HBaseWinProc() {
-		++objInstances;
-		if (!eHeapAddr) {
-			eHeapAddr = HeapCreate(HEAP_CREATE_ENABLE_EXECUTE | HEAP_GENERATE_EXCEPTIONS, 0, 0);
 
-			if (!eHeapAddr)
-				CheckDefaultWinError;
-		}
-		try {
-			thunk = new(eHeapAddr)winThunk;
-		}
-		catch (...) {
-			throw Himani::Exceptions(L"Thunk could not be allocated");
+	class HDialogProc :public FunctionThunk {
+	public:
+		HDialogProc() {
+			thunk->Init(this, (DWORD_PTR)WndProc);
 		}
 
-		if (!thunk->Init(this, (DWORD_PTR)WndProc))
-			CheckDefaultWinError;
-	}
-	HBaseWinProc::~HBaseWinProc() {
-		if (objInstances == 1) {
-			//TOOD Check for Error or not?
-			HeapDestroy(eHeapAddr);
-			eHeapAddr = NULL;
-			objInstances = 0;
+	protected:
+		virtual BOOL MessageFunc(HWND hwnd, UINT message, WPARAM wParam, LPARAM lParam) {
+			return DefWindowProc(hwnd, message, wParam, lParam);
 		}
-		else {
-			objInstances--;
-			delete thunk;
+
+	private:
+#if defined(_M_IX86)
+		static BOOL CALLBACK WndProc(HDialogProc* pThis, HWND hwnd, UINT message, WPARAM wParam, LPARAM lParam) {
+#elif defined(_M_AMD64)
+		static BOOL CALLBACK WndProc(HWND hwnd, UINT message, WPARAM wParam, LPARAM lParam, HDialogProc* pThis) {
+#endif
+			return pThis->MessageFunc(hwnd, message, wParam, lParam);
 		}
-	}
 
-	HANDLE HBaseWinProc::eHeapAddr = nullptr;
-	int HBaseWinProc::objInstances = 0;
-
+	};
 
 #pragma endregion
